Refactored SemEval evaluation experiments and moved all experiment classes to different python scripts under the 'experiments' directory
