+++ eval.py
@@ -1,7 +1,9 @@
 import argparse
 from common import set_custom_seed
+from losses.base import ModelLoader
 from distances import EuclideanDistance, CosineDistance
-from experiments import VoxCeleb1ModelEvaluationExperiment, SemEvalModelEvaluationExperiment
+from experiments.semeval import SemEvalEmbeddingEvaluationExperiment, SemEvalBaselineModelEvaluationExperiment
+from experiments.voxceleb import VoxCeleb1ModelEvaluationExperiment
 
 # Script arguments
 parser = argparse.ArgumentParser()
@@ -20,7 +22,12 @@ args = parser.parse_args()
 # Set custom seed
 set_custom_seed()
 
-distance = CosineDistance() if args.distance == 'cosine' else EuclideanDistance()
+if args.distance == 'cosine':
+    distance = CosineDistance()
+elif args.distance == 'euclidean':
+    distance = EuclideanDistance()
+else:
+    raise ValueError("Distance can only be: cosine / euclidean")
 
 print(f"[Task: {args.task.upper()}]")
 
@@ -32,15 +39,20 @@ if args.task == 'speaker':
                                                     batch_size=args.batch_size)
     metric_name = 'EER'
 elif args.task == 'sts':
-    experiment = SemEvalModelEvaluationExperiment(model_path=args.model,
-                                                  nfeat=500,
-                                                  data_path=args.sts_path,
-                                                  word2vec_path=args.word2vec,
-                                                  vocab_path=args.vocab,
-                                                  distance=distance,
-                                                  log_interval=args.log_interval,
-                                                  batch_size=args.batch_size,
-                                                  baseline=False)
+    model_loader = ModelLoader(args.model)
+    loss_name = model_loader.get_trained_loss()
+    if loss_name == 'kldiv':
+        experiment_type = SemEvalBaselineModelEvaluationExperiment
+    else:
+        experiment_type = SemEvalEmbeddingEvaluationExperiment
+    experiment = experiment_type(model_loader=model_loader,
+                                 nfeat=500,
+                                 data_path=args.sts_path,
+                                 word2vec_path=args.word2vec,
+                                 vocab_path=args.vocab,
+                                 distance=distance,
+                                 log_interval=args.log_interval,
+                                 batch_size=args.batch_size)
     metric_name = 'Spearman'
 else:
     raise ValueError("Task can only be 'speaker' or 'sts'")
@@ -48,8 +60,10 @@ else:
 print('[Started Evaluation...]')
 if args.partition == 'dev':
     metric = experiment.evaluate_on_dev()
-else:
+elif args.partition == 'test':
     metric = experiment.evaluate_on_test()
+else:
+    raise ValueError('Partition can only be: dev / test')
 print(f"[Evaluation Finished]")
 
 print(f"[{args.partition.upper()} {metric_name} = {metric}]")
